refactor(whitenoise): split groups and users into focused submodules
