+++ src/whitenoise/groups.rs
@@ -1,101 +1,29 @@
-use std::{
-    collections::{HashMap, HashSet},
-    path::{Path, PathBuf},
-    time::Duration,
-};
+use std::collections::{HashMap, HashSet};
+use std::time::Duration;
 
-use mdk_core::encrypted_media::types::MediaReference;
-use mdk_core::extension::group_image;
-use mdk_core::media_processing::MediaProcessingOptions;
 use mdk_core::prelude::*;
-use mdk_sqlite_storage::MdkSqliteStorage;
-use mdk_storage_traits::Secret;
-use nostr_blossom::client::BlossomClient;
 use nostr_sdk::prelude::*;
-use sha2::{Digest, Sha256};
 
 use crate::{
     RelayType,
-    types::ImageType,
     whitenoise::{
         Whitenoise,
         accounts::Account,
         accounts_groups::AccountGroup,
-        database::media_files::{FileMetadata, MediaFile},
         error::{Result, WhitenoiseError},
         group_information::{GroupInformation, GroupType},
-        media_files::MediaFileUpload,
         relays::Relay,
         users::User,
     },
 };
 
-/// Default timeout for Blossom HTTP operations (download and upload)
-/// Set to 300 seconds to accommodate large image files over slow connections
-const BLOSSOM_TIMEOUT: Duration = Duration::from_secs(300);
+mod media;
+mod membership;
+mod publish;
 
-/// An MLS group paired with its account-specific membership data.
-///
-/// This struct combines an MLS group with its `AccountGroup` record (the join table
-/// between accounts and groups). The `AccountGroup` tracks account-specific state
-/// such as user confirmation status (pending, accepted, or declined).
-#[derive(Debug, Clone)]
-pub struct GroupWithMembership {
-    /// The MLS group data from MDK
-    pub group: group_types::Group,
-    /// The account-group relationship (join table record)
-    pub membership: AccountGroup,
-}
-
-impl GroupWithMembership {
-    /// Returns true if this group is pending user confirmation.
-    pub fn is_pending(&self) -> bool {
-        self.membership.is_pending()
-    }
-
-    /// Returns true if this group has been accepted by the user.
-    pub fn is_accepted(&self) -> bool {
-        self.membership.is_accepted()
-    }
-}
+pub use membership::GroupWithMembership;
 
 impl Whitenoise {
-    /// Returns the default Blossom server URL based on build configuration
-    ///
-    /// In debug builds, uses localhost:3000 for local testing.
-    /// In release builds, uses the production Blossom server.
-    fn default_blossom_url() -> Url {
-        let url = if cfg!(debug_assertions) {
-            "http://localhost:3000"
-        } else {
-            "https://blossom.primal.net"
-        };
-        Url::parse(url).expect("Hardcoded Blossom URL should be valid")
-    }
-
-    /// Ensures that group relays are available for publishing evolution events.
-    /// Returns the validated relay URLs.
-    ///
-    /// # Arguments
-    /// * `mdk` - The NostrMls instance to get relays from
-    /// * `group_id` - The ID of the group
-    ///
-    /// # Returns
-    /// * `Ok(Vec<nostr_sdk::RelayUrl>)` - Vector of relay URLs
-    /// * `Err(WhitenoiseError::GroupMissingRelays)` - If no relays are configured
-    pub(crate) fn ensure_group_relays(
-        mdk: &MDK<MdkSqliteStorage>,
-        group_id: &GroupId,
-    ) -> Result<Vec<nostr_sdk::RelayUrl>> {
-        let group_relays = mdk.get_relays(group_id)?;
-
-        if group_relays.is_empty() {
-            return Err(WhitenoiseError::GroupMissingRelays);
-        }
-
-        Ok(group_relays.into_iter().collect())
-    }
-
     async fn resolve_member_delivery_relays(
         &self,
         member: &User,
@@ -136,121 +64,6 @@ impl Whitenoise {
         }
 
         Ok(fallback_relays)
-    }
-
-    /// Maximum number of retry attempts when publishing an event to relays.
-    const MAX_PUBLISH_ATTEMPTS: u32 = 3;
-
-    /// Publishes a pre-signed event to relays with retry and exponential backoff.
-    ///
-    /// Attempts to publish up to [`Self::MAX_PUBLISH_ATTEMPTS`] times with
-    /// exponential backoff (2 s, 4 s). The event is created once by the caller;
-    /// only the relay publish step is retried. A publish is considered
-    /// successful only when at least one relay accepts the event.
-    ///
-    /// This is the single entry-point for publishing MLS protocol events
-    /// (evolution commits, proposals, etc.) so that retry policy changes are
-    /// made in one place. When a durable publish queue is introduced later,
-    /// only this method needs to be replaced.
-    pub(crate) async fn publish_event_with_retry(
-        &self,
-        event: Event,
-        account_pubkey: &PublicKey,
-        relay_urls: &[RelayUrl],
-    ) -> Result<()> {
-        let mut last_error = None;
-
-        for attempt in 0..Self::MAX_PUBLISH_ATTEMPTS {
-            if attempt > 0 {
-                let delay = Duration::from_secs(1 << attempt);
-                tracing::warn!(
-                    target: "whitenoise::groups::publish_event_with_retry",
-                    "Retrying event publish (attempt {}/{}), backing off {delay:?}",
-                    attempt + 1,
-                    Self::MAX_PUBLISH_ATTEMPTS,
-                );
-                tokio::time::sleep(delay).await;
-            }
-
-            match self
-                .nostr
-                .publish_event_to(event.clone(), account_pubkey, relay_urls)
-                .await
-            {
-                Ok(output) if !output.success.is_empty() => return Ok(()),
-                Ok(output) => {
-                    // API call succeeded but no relay accepted the event
-                    tracing::warn!(
-                        target: "whitenoise::groups::publish_event_with_retry",
-                        "Event publish attempt {}/{}: no relay accepted \
-                         (failed: {:?})",
-                        attempt + 1,
-                        Self::MAX_PUBLISH_ATTEMPTS,
-                        output.failed.keys().collect::<Vec<_>>(),
-                    );
-                    last_error = Some(WhitenoiseError::EventPublishNoRelayAccepted);
-                }
-                Err(e) => {
-                    tracing::warn!(
-                        target: "whitenoise::groups::publish_event_with_retry",
-                        "Event publish attempt {}/{} failed: {}",
-                        attempt + 1,
-                        Self::MAX_PUBLISH_ATTEMPTS,
-                        e,
-                    );
-                    last_error = Some(e.into());
-                }
-            }
-        }
-
-        Err(last_error.expect("loop ran at least once"))
-    }
-
-    /// Publishes an evolution event and merges the pending commit on success.
-    ///
-    /// Per MIP-03 this is the canonical ordering for MLS state evolution:
-    /// 1. Caller creates the pending commit via an MDK operation
-    /// 2. This method publishes the evolution event (with retry)
-    /// 3. Only after at least one relay accepts, the pending commit is merged
-    ///
-    /// # Publish failure and rollback
-    ///
-    /// If all publish attempts fail, the pending commit is cleared via
-    /// `clear_pending_commit`, rolling back the MLS group to its pre-commit
-    /// state. This ensures a failed publish never leaves the group stuck with
-    /// a dangling pending commit that would block all subsequent operations.
-    /// The error from the publish attempt is returned to the caller.
-    pub(crate) async fn publish_and_merge_commit(
-        &self,
-        evolution_event: Event,
-        account_pubkey: &PublicKey,
-        group_id: &GroupId,
-        relay_urls: &[RelayUrl],
-    ) -> Result<()> {
-        let mdk = self.create_mdk_for_account(*account_pubkey)?;
-
-        if let Err(publish_err) = self
-            .publish_event_with_retry(evolution_event, account_pubkey, relay_urls)
-            .await
-        {
-            // Publish failed — roll back the pending commit so the group is
-            // not left in a blocked state. Log but do not propagate the
-            // clear error; the original publish error is what matters to the caller.
-            if let Err(clear_err) = mdk.clear_pending_commit(group_id) {
-                tracing::warn!(
-                    target: "whitenoise::groups::publish_and_merge_commit",
-                    "Failed to clear pending commit after publish failure for group {}: {}",
-                    hex::encode(group_id.as_slice()),
-                    clear_err,
-                );
-            }
-            return Err(publish_err);
-        }
-
-        // Relay accepted — now safe to advance local MLS state
-        mdk.merge_pending_commit(group_id)?;
-
-        Ok(())
     }
 
     /// Creates a new MLS group with the specified members and settings
@@ -768,1073 +581,6 @@ impl Whitenoise {
         // TODO: Do any local updates to ensure that we're accurately reflecting that the account is trying to leave this group
         Ok(())
     }
-
-    /// Syncs group image cache if needed (smart, hash-based check)
-    ///
-    /// This method is called after processing welcomes and commits to proactively
-    /// cache group images. It only downloads if:
-    /// 1. The group has an image set
-    /// 2. The image_hash is not already cached
-    ///
-    /// This ensures images are ready before the UI needs them, while avoiding
-    /// redundant downloads.
-    ///
-    /// # Arguments
-    /// * `account` - The account viewing the group
-    /// * `group_id` - The MLS group ID
-    pub(crate) async fn sync_group_image_cache_if_needed(
-        &self,
-        account: &Account,
-        group_id: &GroupId,
-    ) -> Result<()> {
-        let group: mdk_core::prelude::group_types::Group;
-        {
-            // Get group data to check if it has an image
-            let mdk = self.create_mdk_for_account(account.pubkey)?;
-            group = mdk
-                .get_group(group_id)
-                .map_err(WhitenoiseError::from)?
-                .ok_or(WhitenoiseError::GroupNotFound)?;
-        }
-
-        // Check if group has an image set
-        let (image_hash, image_key, image_nonce) =
-            match (group.image_hash, group.image_key, group.image_nonce) {
-                (Some(hash), Some(key), Some(nonce)) => (hash, key, nonce),
-                _ => return Ok(()), // No image set, nothing to do
-            };
-
-        // Try to get the stored blossom_url from the database
-        let blossom_url = if let Some(media_file) =
-            crate::whitenoise::database::media_files::MediaFile::find_by_hash(
-                &self.database,
-                &image_hash,
-            )
-            .await?
-        {
-            media_file
-                .blossom_url
-                .and_then(|url_str| Url::parse(&url_str).ok())
-        } else {
-            None
-        };
-
-        // Download and cache the image
-        self.download_and_cache_group_image(
-            blossom_url,
-            &account.pubkey,
-            group_id,
-            &image_hash,
-            &image_key,
-            &image_nonce,
-        )
-        .await?;
-
-        Ok(())
-    }
-
-    /// Spawns a background task to sync group image cache without blocking
-    ///
-    /// This is used by event handlers to proactively cache group images
-    /// without blocking event processing. Failures are logged but don't
-    /// affect the caller - images will download on-demand if needed.
-    ///
-    /// # Arguments
-    /// * `account` - The account viewing the group
-    /// * `group_id` - The MLS group ID
-    pub(crate) fn background_sync_group_image_cache_if_needed(
-        account: &Account,
-        group_id: &GroupId,
-    ) {
-        let account_clone = account.clone();
-        let group_id_clone = group_id.clone();
-        tokio::spawn(async move {
-            let whitenoise = match Whitenoise::get_instance() {
-                Ok(wn) => wn,
-                Err(e) => {
-                    tracing::error!(
-                        target: "whitenoise::groups::background_sync_group_image_cache_if_needed",
-                        "Failed to get Whitenoise instance for background image cache: {}",
-                        e
-                    );
-                    return;
-                }
-            };
-
-            if let Err(e) = whitenoise
-                .sync_group_image_cache_if_needed(&account_clone, &group_id_clone)
-                .await
-            {
-                tracing::warn!(
-                    target: "whitenoise::groups::background_sync_group_image_cache_if_needed",
-                    "Background image cache failed: {}. Image will download on-demand.",
-                    e
-                );
-            }
-        });
-    }
-
-    /// Downloads, decrypts, and caches a group image if not already cached
-    ///
-    /// # Arguments
-    /// * `blossom_url` - Optional Blossom server URL (uses default if None)
-    /// * `account_pubkey` - The account accessing the image
-    /// * `group_id` - The MLS group ID
-    /// * `image_hash` - SHA-256 hash of the encrypted image
-    /// * `image_key` - Encryption key
-    /// * `image_nonce` - Encryption nonce
-    ///
-    /// # Returns
-    /// The MediaFile record for the cached image
-    async fn download_and_cache_group_image(
-        &self,
-        blossom_url: Option<Url>,
-        account_pubkey: &PublicKey,
-        group_id: &GroupId,
-        image_hash: &[u8; 32],
-        image_key: &[u8; 32],
-        image_nonce: &[u8; 12],
-    ) -> Result<MediaFile> {
-        let hash_hex = hex::encode(image_hash);
-
-        // Check if already cached and return early if found
-        if let Some(cached_path) = self.check_cached_image(&hash_hex).await? {
-            let media_file = self
-                .link_cached_image_to_group(
-                    account_pubkey,
-                    group_id,
-                    &cached_path,
-                    image_hash,
-                    image_key,
-                )
-                .await?;
-            return Ok(media_file);
-        }
-
-        // Use provided URL or fall back to default
-        let blossom_url = blossom_url.unwrap_or_else(Self::default_blossom_url);
-
-        tracing::info!(
-            target: "whitenoise::groups::download_and_cache_group_image",
-            "Downloading group image {} for group {} from {}",
-            hash_hex,
-            hex::encode(group_id.as_slice()),
-            blossom_url
-        );
-
-        // Download and decrypt the image (MDK handles hash verification internally)
-        let encrypted_data = Self::download_blob_from_blossom(&blossom_url, image_hash).await?;
-
-        let secret_key = Secret::new(*image_key);
-        let secret_nonce = Secret::new(*image_nonce);
-        let decrypted_data = Self::decrypt_group_image(
-            &encrypted_data,
-            Some(image_hash),
-            &secret_key,
-            &secret_nonce,
-        )?;
-        let image_type = ImageType::detect(&decrypted_data).map_err(|e| {
-            WhitenoiseError::UnsupportedMediaFormat(format!("Failed to detect image type: {}", e))
-        })?;
-
-        tracing::debug!(
-            target: "whitenoise::groups::download_and_cache_group_image",
-            "Detected image type: {} for group image {}",
-            image_type.mime_type(),
-            hash_hex
-        );
-
-        let media_file = self
-            .store_and_record_group_image(
-                account_pubkey,
-                group_id,
-                &decrypted_data,
-                image_hash,
-                image_key,
-                &image_type,
-                &blossom_url,
-            )
-            .await?;
-
-        tracing::info!(
-            target: "whitenoise::groups::download_and_cache_group_image",
-            "Cached group image at: {}",
-            media_file.file_path.display()
-        );
-
-        Ok(media_file)
-    }
-
-    /// Checks if an image is already cached
-    async fn check_cached_image(&self, hash_hex: &str) -> Result<Option<PathBuf>> {
-        let media_files = self.media_files();
-        if let Some(cached_path) = media_files.find_file_with_prefix(hash_hex).await {
-            tracing::debug!(
-                target: "whitenoise::groups::check_cached_image",
-                "Group image already cached at: {}",
-                cached_path.display()
-            );
-            Ok(Some(cached_path))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Links an already cached image to a group in the database
-    ///
-    /// When an image file is already cached on disk (from another group using the same image),
-    /// this method creates a database record linking this group/account to the existing file.
-    ///
-    /// If an existing database record exists for this file hash, it copies metadata from it.
-    /// If no database record exists, it detects the MIME type from the cached file itself.
-    ///
-    /// # Returns
-    /// The MediaFile record that was created or already exists
-    async fn link_cached_image_to_group(
-        &self,
-        account_pubkey: &PublicKey,
-        group_id: &GroupId,
-        cached_path: &Path,
-        image_hash: &[u8; 32],
-        image_key: &[u8; 32],
-    ) -> Result<MediaFile> {
-        // Try to get metadata from any existing record with this hash
-        let existing_record_opt = MediaFile::find_by_hash(&self.database, image_hash).await?;
-
-        if let Some(existing_record) = existing_record_opt {
-            self.link_cached_image_from_existing_record(
-                account_pubkey,
-                group_id,
-                cached_path,
-                image_hash,
-                existing_record,
-            )
-            .await
-        } else {
-            self.link_cached_image_with_detection(
-                account_pubkey,
-                group_id,
-                cached_path,
-                image_hash,
-                image_key,
-            )
-            .await
-        }
-    }
-
-    /// Links a cached image using metadata from an existing database record
-    async fn link_cached_image_from_existing_record(
-        &self,
-        account_pubkey: &PublicKey,
-        group_id: &GroupId,
-        cached_path: &Path,
-        image_hash: &[u8; 32],
-        existing_record: crate::whitenoise::database::media_files::MediaFile,
-    ) -> Result<MediaFile> {
-        let metadata_ref = existing_record.file_metadata.as_ref();
-        let original_hash_ref = existing_record
-            .original_file_hash
-            .as_ref()
-            .and_then(|hash| hash.as_slice().try_into().ok());
-        let upload = MediaFileUpload {
-            data: &[], // Empty is OK - file already exists on disk
-            original_file_hash: original_hash_ref.as_ref(),
-            encrypted_file_hash: *image_hash,
-            mime_type: &existing_record.mime_type,
-            media_type: &existing_record.media_type,
-            blossom_url: existing_record.blossom_url.as_deref(),
-            nostr_key: existing_record.nostr_key.clone(),
-            file_metadata: metadata_ref,
-            nonce: existing_record.nonce.as_deref().map(|s| s.to_string()),
-            scheme_version: existing_record.scheme_version.as_deref(),
-        };
-
-        self.media_files()
-            .record_in_database(account_pubkey, group_id, cached_path, upload)
-            .await
-    }
-
-    /// Links a cached image by detecting its MIME type from the file
-    async fn link_cached_image_with_detection(
-        &self,
-        account_pubkey: &PublicKey,
-        group_id: &GroupId,
-        cached_path: &Path,
-        image_hash: &[u8; 32],
-        image_key: &[u8; 32],
-    ) -> Result<MediaFile> {
-        tracing::debug!(
-            target: "whitenoise::groups::link_cached_image_with_detection",
-            "No existing database record for hash {}, detecting MIME type from cached file",
-            hex::encode(image_hash)
-        );
-
-        let file_data = tokio::fs::read(cached_path).await?;
-
-        let image_type = ImageType::detect(&file_data).map_err(|e| {
-            WhitenoiseError::UnsupportedMediaFormat(format!(
-                "Failed to detect image type for cached file {}: {}",
-                cached_path.display(),
-                e
-            ))
-        })?;
-
-        // Compute original_file_hash from the decrypted data
-        let mut hasher = Sha256::new();
-        hasher.update(&file_data);
-        let original_hash: [u8; 32] = hasher.finalize().into();
-
-        // Derive the upload keypair from the image key for cleanup purposes
-        let secret_key = Secret::new(*image_key);
-        let upload_keypair = group_image::derive_upload_keypair(&secret_key, 2).map_err(|e| {
-            WhitenoiseError::Other(anyhow::anyhow!("Failed to derive upload keypair: {}", e))
-        })?;
-
-        let upload = MediaFileUpload {
-            data: &[], // Empty is OK - file already exists on disk
-            original_file_hash: Some(&original_hash),
-            encrypted_file_hash: *image_hash,
-            mime_type: image_type.mime_type(),
-            media_type: "group_image",
-            blossom_url: None,
-            nostr_key: Some(upload_keypair.secret_key().to_secret_hex()),
-            file_metadata: None,
-            nonce: None, // Group images use key/nonce encryption, not MDK
-            scheme_version: None,
-        };
-
-        self.media_files()
-            .record_in_database(account_pubkey, group_id, cached_path, upload)
-            .await
-    }
-
-    /// Downloads an encrypted blob from a Blossom server
-    async fn download_blob_from_blossom(
-        blossom_url: &Url,
-        image_hash: &[u8; 32],
-    ) -> Result<Vec<u8>> {
-        use nostr::hashes::{Hash, sha256::Hash as Sha256Hash};
-
-        let client = BlossomClient::new(blossom_url.clone());
-        let sha256 = Sha256Hash::from_slice(image_hash)
-            .map_err(|e| WhitenoiseError::Other(anyhow::anyhow!("Invalid SHA256 hash: {}", e)))?;
-
-        let download_future = client.get_blob(sha256, None, None, None::<&Keys>);
-
-        tokio::time::timeout(BLOSSOM_TIMEOUT, download_future)
-            .await
-            .map_err(|_| {
-                WhitenoiseError::BlossomDownload(format!(
-                    "Download timed out after {} seconds",
-                    BLOSSOM_TIMEOUT.as_secs()
-                ))
-            })?
-            .map_err(|e| {
-                WhitenoiseError::BlossomDownload(format!("Failed to download blob: {}", e))
-            })
-    }
-
-    /// Decrypts a group image using the provided key and nonce
-    fn decrypt_group_image(
-        encrypted_data: &[u8],
-        expected_hash: Option<&[u8; 32]>,
-        image_key: &Secret<[u8; 32]>,
-        image_nonce: &Secret<[u8; 12]>,
-    ) -> Result<Vec<u8>> {
-        group_image::decrypt_group_image(encrypted_data, expected_hash, image_key, image_nonce)
-            .map_err(|e| {
-                WhitenoiseError::ImageDecryptionFailed(format!(
-                    "Failed to decrypt group image: {}",
-                    e
-                ))
-            })
-    }
-
-    /// Downloads and decrypts a chat media blob
-    ///
-    /// This helper orchestrates the download from Blossom, hash verification,
-    /// and MDK decryption in a single focused operation.
-    ///
-    /// # Arguments
-    /// * `account_pubkey` - The account downloading the media
-    /// * `data_dir` - The data directory for MDK storage
-    /// * `group_id` - The MLS group ID
-    /// * `media_file` - The MediaFile record containing URLs, hashes, and metadata
-    /// * `original_file_hash` - SHA-256 of original content (for MDK decryption)
-    ///
-    /// # Returns
-    /// * `Ok(Vec<u8>)` - Decrypted file data
-    /// * `Err(WhitenoiseError)` - If download, verification, or decryption fails
-    async fn download_and_decrypt_chat_media_blob(
-        account_pubkey: &PublicKey,
-        data_dir: &Path,
-        keyring_service_id: &str,
-        group_id: &GroupId,
-        media_file: &MediaFile,
-        original_file_hash: &[u8; 32],
-    ) -> Result<Vec<u8>> {
-        // Extract filename required for MDK AAD (cryptographically bound)
-        let filename = media_file
-            .file_metadata
-            .as_ref()
-            .and_then(|m| m.original_filename.as_deref())
-            .ok_or_else(|| {
-                WhitenoiseError::MediaCache("Missing required filename metadata".to_string())
-            })?;
-        // Parse encrypted_file_hash for download verification
-        let encrypted_hash: [u8; 32] = media_file
-            .encrypted_file_hash
-            .as_slice()
-            .try_into()
-            .map_err(|_| {
-                WhitenoiseError::MediaCache("Invalid encrypted_file_hash length".to_string())
-            })?;
-
-        // Parse Blossom URL
-        let blossom_url_str = media_file
-            .blossom_url
-            .as_ref()
-            .ok_or_else(|| WhitenoiseError::MediaCache("No Blossom URL".to_string()))?;
-
-        let blossom_url = Url::parse(blossom_url_str).map_err(|e| {
-            WhitenoiseError::MediaCache(format!("Invalid Blossom URL '{}': {}", blossom_url_str, e))
-        })?;
-
-        tracing::debug!(
-            target: "whitenoise::groups::download_and_decrypt",
-            "Downloading encrypted blob from: {}",
-            blossom_url
-        );
-
-        // Download encrypted blob (includes hash verification)
-        let encrypted_data =
-            Self::download_blob_from_blossom(&blossom_url, &encrypted_hash).await?;
-
-        // Decrypt using MDK
-        let mdk = Account::create_mdk(*account_pubkey, data_dir, keyring_service_id)?;
-        let media_manager = mdk.media_manager(group_id.clone());
-
-        // Retrieve nonce and scheme_version from database (required for MDK decryption)
-        let nonce_hex = media_file.nonce.as_ref().ok_or_else(|| {
-            WhitenoiseError::MediaCache("Missing nonce for chat media".to_string())
-        })?;
-        let scheme_version = media_file
-            .scheme_version
-            .as_ref()
-            .ok_or_else(|| {
-                WhitenoiseError::MediaCache("Missing scheme_version for chat media".to_string())
-            })?
-            .clone();
-
-        // Decode nonce from hex string to bytes
-        let nonce_bytes = hex::decode(nonce_hex)
-            .map_err(|_| WhitenoiseError::MediaCache("Invalid nonce hex".to_string()))?;
-        let nonce: [u8; 12] = nonce_bytes
-            .try_into()
-            .map_err(|_| WhitenoiseError::MediaCache("Invalid nonce length".to_string()))?;
-
-        let reference = MediaReference {
-            url: String::new(),
-            original_hash: *original_file_hash,
-            mime_type: media_file.mime_type.clone(),
-            filename: filename.to_string(),
-            dimensions: None, // Not used in decryption (only display metadata)
-            scheme_version,
-            nonce,
-        };
-
-        media_manager
-            .decrypt_from_download(&encrypted_data, &reference)
-            .map_err(|e| {
-                WhitenoiseError::Other(anyhow::anyhow!("Failed to decrypt chat media: {}", e))
-            })
-    }
-
-    /// Stores decrypted image to cache and records it in database
-    async fn store_and_record_group_image(
-        &self,
-        account_pubkey: &PublicKey,
-        group_id: &GroupId,
-        decrypted_data: &[u8],
-        image_hash: &[u8; 32],
-        image_key: &[u8; 32],
-        image_type: &ImageType,
-        blossom_server: &Url,
-    ) -> Result<MediaFile> {
-        let hash_hex = hex::encode(image_hash);
-        let filename = format!("{}.{}", hash_hex, image_type.extension());
-        let blossom_url = blossom_server.join(&hash_hex).map_err(|e| {
-            WhitenoiseError::Other(anyhow::anyhow!("Failed to construct Blossom URL: {}", e))
-        })?;
-
-        // Compute original_file_hash from the decrypted data
-        let mut hasher = Sha256::new();
-        hasher.update(decrypted_data);
-        let original_hash: [u8; 32] = hasher.finalize().into();
-
-        // Derive the upload keypair from the image key for cleanup purposes
-        let secret_key = Secret::new(*image_key);
-        let upload_keypair = group_image::derive_upload_keypair(&secret_key, 2).map_err(|e| {
-            WhitenoiseError::Other(anyhow::anyhow!("Failed to derive upload keypair: {}", e))
-        })?;
-
-        let upload = MediaFileUpload {
-            data: decrypted_data,
-            original_file_hash: Some(&original_hash),
-            encrypted_file_hash: *image_hash,
-            mime_type: image_type.mime_type(),
-            media_type: "group_image",
-            blossom_url: Some(blossom_url.as_str()),
-            nostr_key: Some(upload_keypair.secret_key().to_secret_hex()),
-            file_metadata: None,
-            nonce: None, // Group images use key/nonce encryption, not MDK
-            scheme_version: None,
-        };
-
-        self.media_files()
-            .store_and_record(account_pubkey, group_id, &filename, upload)
-            .await
-    }
-
-    /// Uploads encrypted blob to Blossom server with timeout
-    ///
-    /// # Arguments
-    /// * `blossom_server_url` - Blossom server URL to upload to
-    /// * `encrypted_data` - The encrypted image data to upload
-    /// * `mime_type` - MIME type of the original image
-    /// * `upload_keypair` - Keypair for signing the upload
-    ///
-    /// # Returns
-    /// * `Ok(BlobDescriptor)` - Descriptor from Blossom server containing URL and hash
-    /// * `Err(WhitenoiseError)` - Upload timeout or network error
-    async fn upload_encrypted_blob_to_blossom(
-        blossom_server_url: &Url,
-        encrypted_data: Vec<u8>,
-        mime_type: &str,
-        upload_keypair: &Keys,
-    ) -> Result<nostr_blossom::bud02::BlobDescriptor> {
-        let client = BlossomClient::new(blossom_server_url.clone());
-        let upload_future = client.upload_blob(
-            encrypted_data,
-            Some(mime_type.to_string()),
-            None,
-            Some(upload_keypair),
-        );
-
-        tokio::time::timeout(BLOSSOM_TIMEOUT, upload_future)
-            .await
-            .map_err(|_| {
-                WhitenoiseError::Other(anyhow::anyhow!(
-                    "Upload timed out after {} seconds",
-                    BLOSSOM_TIMEOUT.as_secs()
-                ))
-            })?
-            .map_err(|err| WhitenoiseError::Other(anyhow::anyhow!(err)))
-    }
-
-    /// Uploads a group image to a Blossom server and returns the encrypted metadata.
-    ///
-    /// The returned metadata (hash, key, nonce) should be passed to `update_group_data`
-    /// to update the group's image settings.
-    ///
-    /// # Arguments
-    /// * `account` - The account performing the upload (must be a group admin)
-    /// * `group_id` - The ID of the group to upload the image for
-    /// * `file_path` - Path to the image file to upload
-    /// * `blossom_server_url` - Blossom server URL to upload to
-    /// * `options` - Optional media processing options (defaults to standard options if None)
-    pub async fn upload_group_image(
-        &self,
-        account: &Account,
-        group_id: &GroupId,
-        file_path: &str,
-        blossom_server_url: Option<Url>,
-        options: Option<MediaProcessingOptions>,
-    ) -> Result<([u8; 32], [u8; 32], [u8; 12])> {
-        // Verify the account is an admin of the group
-        let admins = self.group_admins(account, group_id).await?;
-        if !admins.contains(&account.pubkey) {
-            return Err(WhitenoiseError::AccountNotAuthorized);
-        }
-
-        // Read the image file
-        let image_data = tokio::fs::read(file_path).await?;
-
-        // Detect and validate image type from file content
-        // This uses the image crate to both detect the format and validate the image
-        let image_type = ImageType::detect(&image_data).map_err(|e| {
-            WhitenoiseError::UnsupportedMediaFormat(format!(
-                "Failed to detect or validate image from {}: {}",
-                file_path, e
-            ))
-        })?;
-
-        tracing::debug!(
-            target: "whitenoise::groups::upload_group_image",
-            "Detected and validated image type: {} for file {}",
-            image_type.mime_type(),
-            file_path
-        );
-
-        // Use MDK to prepare the image for upload (encrypt + derive keypair)
-        let prepared = group_image::prepare_group_image_for_upload_with_options(
-            &image_data,
-            image_type.mime_type(),
-            &options.unwrap_or_default(),
-        )
-        .map_err(|e| {
-            WhitenoiseError::Other(anyhow::anyhow!("Failed to prepare group image: {}", e))
-        })?;
-
-        let blossom_server_url = blossom_server_url.unwrap_or(Self::default_blossom_url());
-        // Upload encrypted data to Blossom using the derived keypair
-        // Extract the encrypted data from Secret wrapper for upload
-        let descriptor = Self::upload_encrypted_blob_to_blossom(
-            &blossom_server_url,
-            prepared.encrypted_data.as_ref().clone(),
-            image_type.mime_type(),
-            &prepared.upload_keypair,
-        )
-        .await?;
-
-        // Verify the Blossom server returned the expected hash
-        // The descriptor.sha256 is a Hash type from bitcoin_hashes, convert to byte array
-        let returned_hash_bytes: [u8; 32] = *descriptor.sha256.as_ref();
-
-        if returned_hash_bytes != prepared.encrypted_hash {
-            return Err(WhitenoiseError::HashMismatch {
-                expected: hex::encode(prepared.encrypted_hash),
-                actual: hex::encode(returned_hash_bytes),
-            });
-        }
-
-        tracing::debug!(
-            target: "whitenoise::groups::upload_group_image",
-            "Successfully uploaded group image for group {} to Blossom server. Hash: {}",
-            hex::encode(group_id.as_slice()),
-            hex::encode(prepared.encrypted_hash)
-        );
-
-        // Cache the original unencrypted image immediately to avoid re-downloading
-        // This is an optimization: the uploader already has the image in memory
-        let hash_hex = hex::encode(prepared.encrypted_hash);
-        let filename = format!("{}.{}", hash_hex, image_type.extension());
-
-        // Compute original_file_hash from the original image data
-        let mut hasher = Sha256::new();
-        hasher.update(&image_data);
-        let original_hash: [u8; 32] = hasher.finalize().into();
-
-        let upload = MediaFileUpload {
-            data: &image_data,
-            original_file_hash: Some(&original_hash),
-            encrypted_file_hash: prepared.encrypted_hash,
-            mime_type: image_type.mime_type(),
-            media_type: "group_image",
-            blossom_url: Some(descriptor.url.as_str()),
-            nostr_key: Some(prepared.upload_keypair.secret_key().to_secret_hex()),
-            file_metadata: None,
-            nonce: None, // Group images use key/nonce encryption, not MDK
-            scheme_version: None,
-        };
-
-        if let Err(e) = self
-            .media_files()
-            .store_and_record(&account.pubkey, group_id, &filename, upload)
-            .await
-        {
-            tracing::warn!(
-                target: "whitenoise::groups::upload_group_image",
-                "Failed to cache uploaded group image: {}. Image will be downloaded on next access.",
-                e
-            );
-        }
-
-        // Return the metadata needed for group update
-        // Extract raw bytes from Secret wrappers
-        Ok((
-            prepared.encrypted_hash,
-            *prepared.image_key.as_ref(),
-            *prepared.image_nonce.as_ref(),
-        ))
-    }
-
-    /// Uploads a media file to a Blossom server and returns the media file record.
-    ///
-    /// This method is designed for media files sent in group chat messages. Unlike
-    /// `upload_group_image`, it does not require admin privileges since any group
-    /// member can send media in chat.
-    ///
-    /// Supports images (JPEG, PNG, GIF, WebP), videos (MP4, WebM, MOV), audio
-    /// (MP3, OGG, M4A, WAV), and documents (PDF).
-    ///
-    /// Uses the encrypted media manager which derives encryption keys from the group secret.
-    /// The encryption keys are not returned as they can be re-derived for decryption.
-    ///
-    /// # Arguments
-    /// * `account` - The account uploading the media file
-    /// * `group_id` - The ID of the group where the media will be used
-    /// * `file_path` - Path to the media file to upload
-    /// * `blossom_server_url` - Optional Blossom server URL (uses default if None)
-    /// * `options` - Optional media processing options (defaults to standard options if None)
-    ///
-    /// # Returns
-    /// * `Ok(MediaFile)` - MediaFile record containing file hash and metadata
-    /// * `Err(WhitenoiseError)` - If media validation, upload, or caching fails
-    pub async fn upload_chat_media(
-        &self,
-        account: &Account,
-        group_id: &GroupId,
-        file_path: &str,
-        blossom_server_url: Option<Url>,
-        options: Option<MediaProcessingOptions>,
-    ) -> Result<MediaFile> {
-        // Read the media file
-        let file_data = tokio::fs::read(file_path).await?;
-
-        // Detect and validate media type from file content
-        let media_detection = crate::types::detect_media_type(&file_data)?;
-
-        tracing::debug!(
-            target: "whitenoise::groups::upload_chat_media",
-            "Detected and validated media type: {} for file {}",
-            media_detection.mime_type(),
-            file_path
-        );
-
-        // Extract filename from path for AAD in encryption
-        let original_filename = std::path::Path::new(file_path)
-            .file_name()
-            .and_then(|n| n.to_str())
-            .ok_or_else(|| WhitenoiseError::Other(anyhow::anyhow!("Invalid file path")))?;
-
-        // Use MDK encrypted media manager to prepare the media file for upload
-        // Wrap in a block to ensure MDK and media_manager are dropped before any await points
-        let prepared = {
-            let mdk = self.create_mdk_for_account(account.pubkey)?;
-            let media_manager = mdk.media_manager(group_id.clone());
-
-            media_manager
-                .encrypt_for_upload_with_options(
-                    &file_data,
-                    media_detection.mime_type(),
-                    original_filename,
-                    &options.unwrap_or_default(),
-                )
-                .map_err(|e| {
-                    WhitenoiseError::Other(anyhow::anyhow!("Failed to encrypt chat media: {}", e))
-                })?
-        };
-
-        let blossom_server_url = blossom_server_url.unwrap_or_else(Self::default_blossom_url);
-
-        // Generate fresh keys for upload authentication (for MIP-04 cleanup)
-        let upload_keys = nostr_sdk::Keys::generate();
-        let upload_keys_hex = upload_keys.secret_key().to_secret_hex();
-
-        // Upload encrypted data to Blossom
-        let descriptor = Self::upload_encrypted_blob_to_blossom(
-            &blossom_server_url,
-            prepared.encrypted_data,
-            &prepared.mime_type,
-            &upload_keys,
-        )
-        .await?;
-
-        // Verify the Blossom server returned the expected hash
-        let returned_hash_bytes: [u8; 32] = *descriptor.sha256.as_ref();
-        if returned_hash_bytes != prepared.encrypted_hash {
-            return Err(WhitenoiseError::HashMismatch {
-                expected: hex::encode(prepared.encrypted_hash),
-                actual: hex::encode(returned_hash_bytes),
-            });
-        }
-
-        tracing::debug!(
-            target: "whitenoise::groups::upload_chat_media",
-            "Successfully uploaded chat media for group {} to Blossom server. Hash: {}",
-            hex::encode(group_id.as_slice()),
-            hex::encode(prepared.encrypted_hash)
-        );
-
-        // Cache the decrypted media file locally
-        let hash_hex = hex::encode(prepared.encrypted_hash);
-        let cached_filename = format!("{}.{}", hash_hex, media_detection.extension());
-
-        // Construct file metadata from the prepared media data
-        let file_metadata = if prepared.dimensions.is_some() || prepared.blurhash.is_some() {
-            Some(FileMetadata {
-                original_filename: Some(prepared.filename.clone()),
-                dimensions: prepared.dimensions.map(|(w, h)| format!("{}x{}", w, h)),
-                blurhash: prepared.blurhash.clone(),
-            })
-        } else {
-            None
-        };
-
-        let upload = MediaFileUpload {
-            data: &file_data,
-            original_file_hash: Some(&prepared.original_hash),
-            encrypted_file_hash: prepared.encrypted_hash,
-            mime_type: &prepared.mime_type,
-            media_type: "chat_media",
-            blossom_url: Some(descriptor.url.as_str()),
-            nostr_key: Some(upload_keys_hex),
-            file_metadata: file_metadata.as_ref(),
-            nonce: Some(hex::encode(prepared.nonce)),
-            scheme_version: Some("mip04-v2"),
-        };
-
-        let media_file = self
-            .media_files()
-            .store_and_record(&account.pubkey, group_id, &cached_filename, upload)
-            .await?;
-
-        Ok(media_file)
-    }
-
-    /// Downloads a chat media file and returns the updated MediaFile record
-    ///
-    /// This method downloads and decrypts media files sent in group chat messages.
-    /// It uses the original file hash from the imeta 'x' field per MIP-04 specification.
-    ///
-    /// **On-demand download**: If the file is already cached, returns immediately.
-    /// **Idempotent**: Safe to call multiple times.
-    ///
-    /// # Arguments
-    /// * `account` - The account downloading the media
-    /// * `group_id` - The MLS group ID where the media was shared
-    /// * `original_file_hash` - The SHA-256 hash of the original file (from imeta 'x' field)
-    ///
-    /// # Returns
-    /// * `Ok(MediaFile)` - MediaFile record with the local file path
-    /// * `Err(WhitenoiseError)` - If download, decryption, or database operation fails
-    ///
-    /// # Errors
-    /// * `MediaCache("MediaFile not found")` - No record exists for this hash in this group
-    /// * `MediaCache("Not chat media")` - The MediaFile is not a chat_media type
-    /// * `MediaCache("Missing required metadata")` - Required filename/dimensions are missing
-    /// * `MediaCache("No Blossom URL")` - MediaFile has no Blossom download URL
-    /// * `HashMismatch` - Downloaded blob doesn't match expected hash
-    /// * Network errors during download
-    /// * MDK decryption errors
-    ///
-    /// # Example
-    /// ```ignore
-    /// // Extract hash from imeta tag's 'x' field
-    /// let original_hash: [u8; 32] = hex::decode(imeta_x_field)?.try_into()?;
-    /// let media_file = whitenoise
-    ///     .download_chat_media(&account, &group_id, &original_hash)
-    ///     .await?;
-    /// display_media(&media_file.file_path);
-    /// ```
-    pub async fn download_chat_media(
-        &self,
-        account: &Account,
-        group_id: &GroupId,
-        original_file_hash: &[u8; 32],
-    ) -> Result<MediaFile> {
-        // Find MediaFile record by original_file_hash + group + account (MIP-04 compliant)
-        let media_file = MediaFile::find_by_original_hash_and_group(
-            &self.database,
-            original_file_hash,
-            group_id,
-            &account.pubkey,
-        )
-        .await?
-        .ok_or_else(|| {
-            WhitenoiseError::MediaCache(format!(
-                "MediaFile not found for original_hash={} in group for account {}",
-                hex::encode(original_file_hash),
-                account.pubkey.to_hex()
-            ))
-        })?;
-
-        // Check if already cached (early return for idempotency)
-        if !media_file.file_path.as_os_str().is_empty() && media_file.file_path.exists() {
-            tracing::debug!(
-                target: "whitenoise::groups::download_chat_media",
-                "Media file already cached at: {}",
-                media_file.file_path.display()
-            );
-            return Ok(media_file);
-        }
-
-        // Verify this is chat_media (not group_image or other types)
-        if media_file.media_type != "chat_media" {
-            return Err(WhitenoiseError::MediaCache(format!(
-                "Not chat media: media_type={}",
-                media_file.media_type
-            )));
-        }
-
-        // Download and decrypt the media blob
-        let decrypted_data = Self::download_and_decrypt_chat_media_blob(
-            &account.pubkey,
-            &self.config.data_dir,
-            &self.config.keyring_service_id,
-            group_id,
-            &media_file,
-            original_file_hash,
-        )
-        .await?;
-
-        // Detect MIME type and extension from decrypted content
-        let media_detection = crate::types::detect_media_type(&decrypted_data)?;
-
-        tracing::debug!(
-            target: "whitenoise::groups::download_chat_media",
-            "Decrypted media: {} ({})",
-            media_detection.mime_type(),
-            media_detection.extension()
-        );
-
-        // Store decrypted file to cache
-        let hash_hex = hex::encode(&media_file.encrypted_file_hash);
-        let cached_filename = format!("{}.{}", hash_hex, media_detection.extension());
-
-        let cache_path = self
-            .storage
-            .media_files
-            .store_file(&cached_filename, &decrypted_data)
-            .await?;
-
-        tracing::debug!(
-            target: "whitenoise::groups::download_chat_media",
-            "Cached to: {}",
-            cache_path.display()
-        );
-
-        // Update database record with file path
-        let media_file_id = media_file.id.ok_or_else(|| {
-            WhitenoiseError::MediaCache("MediaFile record missing id".to_string())
-        })?;
-
-        let updated_file =
-            MediaFile::update_file_path(&self.database, media_file_id, &cache_path).await?;
-
-        Ok(updated_file)
-    }
-
-    /// Retrieves all media files for a specific group
-    ///
-    /// Returns all MediaFile records associated with the group, including:
-    /// - Chat media uploaded by any group member
-    /// - Group images
-    /// - Media references from received messages (may have empty file_path if not downloaded)
-    ///
-    /// # Arguments
-    /// * `group_id` - The MLS group ID to fetch media files for
-    ///
-    /// # Returns
-    /// * `Ok(Vec<MediaFile>)` - List of all media files for the group
-    /// * `Err(WhitenoiseError)` - If database query fails
-    ///
-    /// # Example
-    /// ```ignore
-    /// let media_files = whitenoise.get_media_files_for_group(&group_id).await?;
-    /// for file in media_files {
-    ///     println!("Media: {} ({})", file.mime_type, file.media_type);
-    /// }
-    /// ```
-    pub async fn get_media_files_for_group(&self, group_id: &GroupId) -> Result<Vec<MediaFile>> {
-        MediaFile::find_by_group(&self.database, group_id).await
-    }
-
-    /// Gets the local file path for a group's current image
-    ///
-    /// This is the primary method for UI/Flutter to retrieve group images.
-    /// If the image is already cached, returns the path immediately.
-    /// If not cached, downloads, decrypts, and caches it first.
-    ///
-    /// **This is an on-demand method** - it only downloads when the UI actually
-    /// needs to display the image, avoiding unnecessary network traffic.
-    ///
-    /// # Arguments
-    /// * `account` - The account viewing the group
-    /// * `group_id` - The MLS group ID
-    ///
-    /// # Returns
-    /// * `Ok(Some(PathBuf))` - Path to cached decrypted image
-    /// * `Ok(None)` - Group has no image set
-    /// * `Err(WhitenoiseError)` - Download or cache operation failed
-    ///
-    /// # Example
-    /// ```ignore
-    /// // In UI code - only downloads if not already cached
-    /// if let Some(image_path) = whitenoise.get_group_image_path(&account, &group_id).await? {
-    ///     display_image(image_path);
-    /// }
-    /// ```
-    pub async fn get_group_image_path(
-        &self,
-        account: &Account,
-        group_id: &GroupId,
-    ) -> Result<Option<PathBuf>> {
-        let mdk = self.create_mdk_for_account(account.pubkey)?;
-        let group = mdk
-            .get_group(group_id)
-            .map_err(WhitenoiseError::from)?
-            .ok_or(WhitenoiseError::GroupNotFound)?;
-
-        self.resolve_group_image_path(account, &group).await
-    }
-
-    /// Resolves the cached image path for a group (core logic).
-    ///
-    /// This is the internal implementation that receives the Group directly,
-    /// avoiding redundant MDK instantiation and group fetching when the caller
-    /// already has the group data.
-    ///
-    /// # Arguments
-    /// * `account` - The account accessing the image
-    /// * `group` - The group to resolve the image for
-    ///
-    /// # Returns
-    /// * `Ok(Some(PathBuf))` - Path to cached decrypted image
-    /// * `Ok(None)` - Group has no image configured
-    /// * `Err(...)` - Download or cache operation failed
-    pub(crate) async fn resolve_group_image_path(
-        &self,
-        account: &Account,
-        group: &group_types::Group,
-    ) -> Result<Option<PathBuf>> {
-        // Check if group has an image set
-        let (image_hash, image_key, image_nonce) =
-            match (&group.image_hash, &group.image_key, &group.image_nonce) {
-                (Some(hash), Some(key), Some(nonce)) => (hash, key, nonce),
-                _ => return Ok(None), // No image set
-            };
-
-        // Try to get the stored blossom_url from the database
-        let blossom_url = if let Some(media_file) =
-            crate::whitenoise::database::media_files::MediaFile::find_by_hash(
-                &self.database,
-                image_hash,
-            )
-            .await?
-        {
-            media_file
-                .blossom_url
-                .and_then(|url_str| Url::parse(&url_str).ok())
-        } else {
-            None
-        };
-
-        // Download and cache the image (if not already cached)
-        let media_file = self
-            .download_and_cache_group_image(
-                blossom_url,
-                &account.pubkey,
-                &group.mls_group_id,
-                image_hash,
-                image_key.as_ref(),
-                image_nonce.as_ref(),
-            )
-            .await?;
-
-        Ok(Some(media_file.file_path))
-    }
 }
 
 #[cfg(test)]
@@ -1842,6 +588,8 @@ mod tests {
     use super::*;
     use crate::whitenoise::Whitenoise;
     use crate::whitenoise::test_utils::*;
+    use mdk_core::media_processing::MediaProcessingOptions;
+    use mdk_storage_traits::Secret;
     use nostr_sdk::RelayUrl;
 
     #[tokio::test]
