+++ src/whitenoise/users.rs
@@ -1,34 +1,26 @@
-use std::collections::HashSet;
-
 use chrono::{DateTime, Duration, Utc};
 use nostr_sdk::prelude::*;
 use serde::{Deserialize, Serialize};
 
-use crate::{
-    nostr_manager::NostrManager,
-    whitenoise::{
-        Whitenoise,
-        database::processed_events::ProcessedEvent,
-        error::{Result, WhitenoiseError},
-        relays::{Relay, RelayType},
-        utils::timestamp_to_datetime,
-    },
+use crate::whitenoise::{
+    Whitenoise,
+    database::processed_events::ProcessedEvent,
+    error::{Result, WhitenoiseError},
+    relays::{Relay, RelayType},
+    utils::timestamp_to_datetime,
 };
+
+mod key_package;
+mod relay_sync;
+
+pub use key_package::KeyPackageStatus;
+
+#[cfg(test)]
+use key_package::{classify_key_package, has_valid_encoding_tag};
 
 /// TTL for user metadata before it's considered stale and needs refreshing
 /// Set to 24 hours - metadata doesn't change frequently for most users
 const METADATA_TTL_HOURS: i64 = 24;
-
-/// Status of a user's key package on relays.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum KeyPackageStatus {
-    /// User has a valid, compatible key package.
-    Valid(Box<Event>),
-    /// No key package event found on relays.
-    NotFound,
-    /// Key package found but is in an old/incompatible format.
-    Incompatible,
-}
 
 /// Specifies how user metadata and relay lists should be synchronized when finding or creating a user.
 ///
@@ -137,367 +129,12 @@ impl User {
         Ok(())
     }
 
-    /// Fetches the user's MLS key package event from their configured key package relays.
-    ///
-    /// This method retrieves the user's published MLS (Message Layer Security) key package
-    /// from the Nostr network. Key packages are cryptographic objects that contain the user's
-    /// public keys and credentials needed to add them to MLS group conversations.
-    ///
-    /// The method first retrieves the user's key package relay list (NIP-65 kind 10051 events),
-    /// then fetches the most recent MLS key package event (kind 443) from those relays.
-    ///
-    /// # Arguments
-    ///
-    /// * `whitenoise` - The Whitenoise instance used to access the Nostr client and database
-    pub async fn key_package_event(&self, whitenoise: &Whitenoise) -> Result<Option<Event>> {
-        let key_package_relays = self
-            .relays(RelayType::KeyPackage, &whitenoise.database)
-            .await?;
-        let mut key_package_relays_urls_set: HashSet<RelayUrl> =
-            Relay::urls(&key_package_relays).into_iter().collect();
-        if key_package_relays.is_empty() {
-            tracing::warn!(
-                target: "whitenoise::users::key_package_event",
-                "User {} has no key package relays, using nip65 relays",
-                self.pubkey
-            );
-            key_package_relays_urls_set.extend(Relay::urls(
-                &self.relays(RelayType::Nip65, &whitenoise.database).await?,
-            ));
-        }
-        if key_package_relays_urls_set.is_empty() {
-            tracing::warn!(
-                target: "whitenoise::users::key_package_event",
-                "User {} has neither key package nor NIP-65 relays; returning None",
-                self.pubkey
-            );
-            return Ok(None);
-        }
-
-        let key_package_relays_urls: Vec<RelayUrl> =
-            key_package_relays_urls_set.into_iter().collect();
-        let key_package_event = whitenoise
-            .nostr
-            .fetch_user_key_package(self.pubkey, &key_package_relays_urls)
-            .await?;
-        Ok(key_package_event)
-    }
-
-    /// Checks the status of a user's key package on relays.
-    ///
-    /// Similar to [`key_package_event`](Self::key_package_event), but returns a
-    /// [`KeyPackageStatus`] that distinguishes between valid, missing, and incompatible.
-    pub async fn key_package_status(&self, whitenoise: &Whitenoise) -> Result<KeyPackageStatus> {
-        let event = self.key_package_event(whitenoise).await?;
-        Ok(classify_key_package(event))
-    }
-
     pub async fn relays_by_type(
         &self,
         relay_type: RelayType,
         whitenoise: &Whitenoise,
     ) -> Result<Vec<Relay>> {
         self.relays(relay_type, &whitenoise.database).await
-    }
-
-    /// Fetches the latest relay lists for this user from Nostr and updates the local database
-    pub(crate) async fn update_relay_lists(&self, whitenoise: &Whitenoise) -> Result<()> {
-        let initial_query_relays = self.get_query_relays(whitenoise).await?;
-
-        tracing::info!(
-            target: "whitenoise::users::update_relay_lists",
-            "Updating relay lists for user {} using {} query relays",
-            self.pubkey,
-            initial_query_relays.len()
-        );
-
-        let updated_query_relays = self
-            .update_nip65_relays(whitenoise, &initial_query_relays)
-            .await?;
-
-        self.update_secondary_relay_types(whitenoise, &updated_query_relays)
-            .await?;
-
-        tracing::info!(
-            target: "whitenoise::users::update_relay_lists",
-            "Successfully completed relay list updates for user {}",
-            self.pubkey
-        );
-
-        Ok(())
-    }
-
-    async fn get_query_relays(&self, whitenoise: &Whitenoise) -> Result<Vec<Relay>> {
-        let stored_relays = self.relays(RelayType::Nip65, &whitenoise.database).await?;
-
-        if stored_relays.is_empty() {
-            tracing::debug!(
-                target: "whitenoise::users::get_query_relays",
-                "User {} has no stored NIP-65 relays, using fallback relays",
-                self.pubkey,
-            );
-            let urls = whitenoise.fallback_relay_urls().await;
-            Ok(urls.into_iter().map(|url| Relay::new(&url)).collect())
-        } else {
-            Ok(stored_relays)
-        }
-    }
-
-    async fn update_nip65_relays(
-        &self,
-        whitenoise: &Whitenoise,
-        query_relays: &[Relay],
-    ) -> Result<Vec<Relay>> {
-        match self
-            .sync_relays_for_type(whitenoise, RelayType::Nip65, query_relays)
-            .await
-        {
-            Ok(true) => {
-                let refreshed_relays = self.relays(RelayType::Nip65, &whitenoise.database).await?;
-                tracing::info!(
-                    target: "whitenoise::users::update_nip65_relays",
-                    "Updated NIP-65 relays for user {}, now using {} relays for other types",
-                    self.pubkey,
-                    refreshed_relays.len()
-                );
-                Ok(refreshed_relays)
-            }
-            Ok(false) => {
-                tracing::debug!(
-                    target: "whitenoise::users::update_nip65_relays",
-                    "NIP-65 relays unchanged for user {}",
-                    self.pubkey
-                );
-                Ok(query_relays.to_vec())
-            }
-            Err(e) => {
-                tracing::warn!(
-                    target: "whitenoise::users::update_nip65_relays",
-                    "Failed to update NIP-65 relays for user {}: {}, continuing with original relays",
-                    self.pubkey,
-                    e
-                );
-                Ok(query_relays.to_vec())
-            }
-        }
-    }
-
-    async fn update_secondary_relay_types(
-        &self,
-        whitenoise: &Whitenoise,
-        query_relays: &[Relay],
-    ) -> Result<()> {
-        const SECONDARY_RELAY_TYPES: &[RelayType] = &[RelayType::Inbox, RelayType::KeyPackage];
-
-        for &relay_type in SECONDARY_RELAY_TYPES {
-            if let Err(e) = self
-                .sync_relays_for_type(whitenoise, relay_type, query_relays)
-                .await
-            {
-                tracing::warn!(
-                    target: "whitenoise::users::update_secondary_relay_types",
-                    "Failed to update {:?} relays for user {}: {}",
-                    relay_type,
-                    self.pubkey,
-                    e
-                );
-                // Continue with other relay types - individual failures shouldn't stop the process
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Synchronizes stored relays with a new set of relay URLs
-    ///
-    /// This method compares the currently stored relays with a new set of relay URLs,
-    /// removing relays that are no longer present and adding new ones. This is the
-    /// core synchronization logic used by both network-fetched updates and direct
-    /// event processing.
-    ///
-    /// Returns `true` if changes were made, `false` if no changes needed
-    pub(crate) async fn sync_relay_urls(
-        &self,
-        whitenoise: &Whitenoise,
-        relay_type: RelayType,
-        new_relay_urls: &HashSet<RelayUrl>,
-        event_created_at: Option<DateTime<Utc>>,
-    ) -> Result<bool> {
-        // First, check if we should process this event based on timestamp
-        if let Some(new_timestamp) = event_created_at {
-            let newest_stored_timestamp = ProcessedEvent::newest_relay_event_timestamp(
-                &self.pubkey,
-                relay_type,
-                &whitenoise.database,
-            )
-            .await?;
-
-            match newest_stored_timestamp {
-                Some(stored_timestamp)
-                    if new_timestamp.timestamp_millis() <= stored_timestamp.timestamp_millis() =>
-                {
-                    tracing::debug!(
-                        target: "whitenoise::users::sync_relay_urls",
-                        "Ignoring stale {:?} relay event for user {} (event: {}, stored: {})",
-                        relay_type,
-                        self.pubkey,
-                        new_timestamp.timestamp_millis(),
-                        stored_timestamp.timestamp_millis()
-                    );
-                    return Ok(false);
-                }
-                None => {
-                    tracing::debug!(
-                        target: "whitenoise::users::sync_relay_urls",
-                        "No stored {:?} relay timestamps for user {}, accepting new event",
-                        relay_type,
-                        self.pubkey
-                    );
-                }
-                Some(_) => {
-                    tracing::debug!(
-                        target: "whitenoise::users::sync_relay_urls",
-                        "New {:?} relay event is newer for user {}, proceeding with sync",
-                        relay_type,
-                        self.pubkey
-                    );
-                }
-            }
-        }
-
-        let stored_relays = self.relays(relay_type, &whitenoise.database).await?;
-        let stored_urls: HashSet<&RelayUrl> = stored_relays.iter().map(|r| &r.url).collect();
-        let new_urls_set: HashSet<&RelayUrl> = new_relay_urls.iter().collect();
-
-        if stored_urls == new_urls_set {
-            tracing::debug!(
-                target: "whitenoise::users::sync_relay_urls",
-                "No changes needed for {:?} relays for user {}",
-                relay_type,
-                self.pubkey
-            );
-            return Ok(false);
-        }
-
-        // Apply changes
-        tracing::info!(
-            target: "whitenoise::users::sync_relay_urls",
-            "Updating {:?} relays for user {}: {} existing -> {} new",
-            relay_type,
-            self.pubkey,
-            stored_urls.len(),
-            new_urls_set.len()
-        );
-
-        // Remove relays that are no longer needed
-        for existing_relay in &stored_relays {
-            if !new_urls_set.contains(&existing_relay.url)
-                && let Err(e) = self
-                    .remove_relay(existing_relay, relay_type, &whitenoise.database)
-                    .await
-            {
-                tracing::warn!(
-                    target: "whitenoise::users::sync_relay_urls",
-                    "Failed to remove {:?} relay {} for user {}: {}",
-                    relay_type,
-                    existing_relay.url,
-                    self.pubkey,
-                    e
-                );
-            }
-        }
-
-        // Add new relays
-        for new_relay_url in new_relay_urls {
-            if !stored_urls.contains(new_relay_url) {
-                let new_relay = whitenoise
-                    .find_or_create_relay_by_url(new_relay_url)
-                    .await?;
-                if let Err(e) = self
-                    .add_relay(&new_relay, relay_type, &whitenoise.database)
-                    .await
-                {
-                    tracing::warn!(
-                        target: "whitenoise::users::sync_relay_urls",
-                        "Failed to add {:?} relay {} for user {}: {}",
-                        relay_type,
-                        new_relay_url,
-                        self.pubkey,
-                        e
-                    );
-                }
-            }
-        }
-
-        Ok(true)
-    }
-
-    /// Synchronizes relays for a specific type with the network state
-    ///
-    /// Returns `true` if changes were made, `false` if no changes needed
-    async fn sync_relays_for_type(
-        &self,
-        whitenoise: &Whitenoise,
-        relay_type: RelayType,
-        query_relays: &[Relay],
-    ) -> Result<bool> {
-        let relays_urls: Vec<_> = Relay::urls(query_relays);
-        let relay_event = whitenoise
-            .nostr
-            .fetch_user_relays(self.pubkey, relay_type, &relays_urls)
-            .await
-            .map_err(|e| {
-                tracing::warn!(
-                    target: "whitenoise::users::sync_relays_for_type",
-                    "Failed to fetch {:?} relays for user {}: {}",
-                    relay_type, self.pubkey, e
-                );
-                e
-            })?;
-
-        match relay_event {
-            Some(event) => {
-                let relay_hashset: HashSet<_> = NostrManager::relay_urls_from_event(&event);
-                let changed = self
-                    .sync_relay_urls(
-                        whitenoise,
-                        relay_type,
-                        &relay_hashset,
-                        Some(timestamp_to_datetime(event.created_at)?),
-                    )
-                    .await?;
-
-                if changed {
-                    whitenoise
-                        .nostr
-                        .event_tracker
-                        .track_processed_global_event(&event)
-                        .await?;
-
-                    tracing::debug!(
-                        target: "whitenoise::users::sync_relays_for_type",
-                        "Updated {:?} relays for user {} via background sync with event {}",
-                        relay_type, self.pubkey, event.id.to_hex()
-                    );
-                }
-
-                Ok(changed)
-            }
-            None => {
-                tracing::debug!(
-                    target: "whitenoise::users::sync_relays_for_type",
-                    "No {:?} relay events found for user {}",
-                    relay_type, self.pubkey
-                );
-                Ok(false)
-            }
-        }
-    }
-
-    pub(crate) async fn all_users_with_relay_urls(
-        whitenoise: &Whitenoise,
-    ) -> Result<Vec<(PublicKey, Vec<RelayUrl>)>> {
-        Self::all_with_nip65_relay_urls(&whitenoise.database).await
     }
 
     /// Determines whether metadata should be updated based on comprehensive event processing logic.
@@ -591,32 +228,6 @@ impl User {
         };
 
         Ok(should_update)
-    }
-}
-
-/// Checks whether a key package event has the required `["encoding", "base64"]` tag.
-///
-/// Per MIP-00/MIP-02, key package events must include an explicit encoding tag.
-/// Old key packages published before this requirement lack this tag and are
-/// incompatible with current clients.
-fn has_valid_encoding_tag(event: &Event) -> bool {
-    event.tags.iter().any(|tag| {
-        let s = tag.as_slice();
-        s.len() >= 2 && s[0] == "encoding" && s[1] == "base64"
-    })
-}
-
-/// Determines [`KeyPackageStatus`] from an optional key package event.
-fn classify_key_package(event: Option<Event>) -> KeyPackageStatus {
-    match event {
-        None => KeyPackageStatus::NotFound,
-        Some(event) => {
-            if has_valid_encoding_tag(&event) {
-                KeyPackageStatus::Valid(Box::new(event))
-            } else {
-                KeyPackageStatus::Incompatible
-            }
-        }
     }
 }
